Group Policy Model - Part 1
This is the first in a series of patches which implement Group Policy. It
includes the Resource Model, API, and DB implementation for Policy
Classifier and the Policy Action resources.

In the context of the larger Group Policy Model, these two resources are
referenced by a Policy Rule resource which will be introduced in a subsequent
patch.

Gerrit Spec: https://review.openstack.org/#/c/89469
("classifiers" and "actions" resources in the above spec have been renamed to
"policy_classifiers" and "policy_actions")

Implements: blueprint group-based-policy-abstraction

Change-Id: I80358c063d5427aab8f7f94fb06753328fdd8fa9
Author:    Sumit Naiksatam <sumitnaiksatam@gmail.com>
Author:    Bob Kukura <kukura@noironetworks.com>
Author:    Stephen Wong <s3wong@midokura.com>
Author:    Mohammad Banikazemi <mb@us.ibm.com>
Author:    Mandeep Dhami <dhami@noironetworks.com>