Group Policy Model - Part 1
This is the first in a series of patches which implement Group Policy. It
includes Group Policy API, and Resource Model with DB implementation. This
patch contains the resources which are required in the Group Policy Model.
Follow up patches will have the mapping to the Neutron model, and also the
optional resources.

Gerrit Spec: https://review.openstack.org/#/c/89469

Implements: blueprint group-based-policy-abstraction

Change-Id: I80358c063d5427aab8f7f94fb06753328fdd8fa9
Author:    Sumit Naiksatam <sumitnaiksatam@gmail.com>
Author:    Bob Kukura <kukura@noironetworks.com>
Author:    Stephen Wong <s3wong@midokura.com>
Author:    Mohammad Banikazemi <mb@us.ibm.com>
Author:    Mandeep Dhami <dhami@noironetworks.com>