Group Policy API-1: EP, EPG, L2 Policy, L3 Policy
(Path series identifier: GP-API-1)

This is the first in a series of patches which implement Group Policy. It
includes the Resource Model and API for Endpoint, Endpoint Group, L2
Policy, and L3 Policy resources.

In the context of the larger Group Policy Model, the Endpoint Group resource
references Contracts, which will be introduced in a subsequent patch.

The DB and Plugin layers to back these resources will also be introduced in
subsequent patches.

DocImpact

Gerrit Spec: https://review.openstack.org/#/c/89469

Partially-implements: blueprint group-based-policy-abstraction

Change-Id: Ie1cb58b5e3e0112e52a8b4a8e444b9868d9fcc6e
Author:    Sumit Naiksatam <sumitnaiksatam@gmail.com>
Co-Authored-By:    Bob Kukura <kukura@noironetworks.com>
Co-Authored-By:    Stephen Wong <s3wong@midokura.com>
Co-Authored-By:    Mohammad Banikazemi <mb@us.ibm.com>
Co-Authored-By:    Mandeep Dhami <dhami@noironetworks.com>