Group Policy API-2: Classifiers, Actions, and Rules
This is the second API patch in the Group Policy series. It introduces the
Resource Model and API for Policy Classifiers, Policy Actions, and Policy
Rules.

In the context of the larger Group Policy Model, the Policy Rule resource is
referenced by Contracts, which be will introduced in a subsequent patch.

The DB and Plugin layers to back these resources will also be introduced in
subsequent patches.

Gerrit Spec: https://review.openstack.org/#/c/89469
("classifiers" and "actions" resources in the above spec have been renamed to
"policy_classifiers" and "policy_actions")

Partially-implements: blueprint group-based-policy-abstraction

Change-Id: Idf91331573bcf6dea193c222a8aa9b96708d81c8
Author: Sumit Naiksatam <sumitnaiksatam@gmail.com>
Co-Authored-By: Bob Kukura <kukura@noironetworks.com>
Co-Authored-By: Stephen Wong <s3wong@midokura.com>
Co-Authored-By: Mohammad Banikazemi <mb@us.ibm.com>
Co-Authored-By: Mandeep Dhami <dhami@noironetworks.com>